Fix unit test

+++ server/interceptors_test.go
@@ -20,14 +20,14 @@ func TestGetRespErrorInfo(t *testing.T) {
 			"err is not nil",
 			pb.CommonCoinsResponse{Code: 2, Msg: "this will not be returned"},
 			errors.New("this is an error"),
-			uint32(pb.ErrorCode_ERROR_DEFAULT),
+			int64(pb.ErrorCode_ERROR_DEFAULT),
 			"this is an error",
 		},
 		{
 			"no error and resp is nil",
 			nil,
 			nil,
-			uint32(pb.ErrorCode_ERROR_OK),
+			int64(pb.ErrorCode_ERROR_OK),
 			"",
 		},
 		{
@@ -41,7 +41,7 @@ func TestGetRespErrorInfo(t *testing.T) {
 			"no code and msg field",
 			struct{}{},
 			nil,
-			uint32(pb.ErrorCode_ERROR_OK),
+			int64(pb.ErrorCode_ERROR_OK),
 			"",
 		},
 	}
